arrange code in directories
