+++ controller/src/schemas/dataset.py
@@ -1,0 +1,27 @@
+# Copyright 2023 Iguazio
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import List, Optional
+
+from controller.src.schemas.base import BaseWithVerMetadata
+
+
+class Dataset(BaseWithVerMetadata):
+    _top_level_fields = ["task"]
+
+    project_id: Optional[str] = None
+    task: str
+    sources: Optional[List[str]] = None
+    path: str
+    producer: Optional[str] = None
